refactor: improve Kotlin binding build configuration
- Replace Spotless with ktlint plugin for Kotlin-only formatting
- Extract native build tasks to buildSrc/NativeBuildTasks.kt
- Update library name from gluesql_java to gluesql_kotlin
- Remove java-library plugin (Kotlin-only project)
- Update POM metadata (name, description, inception year 2025)
- Configure ktlint to exclude generated UniFFI files
